refactor(KEN-1135): the page sheds its follow effect [no-changelog]
Rebasing onto KEN-1136 put the marketplace detail page at 257 lines,
seven over the ui/*.tsx ratchet, with no baseline row. The seam is the
effect that follows a marketplace elsewhere when the place the page
opened as is unsubscribed: a behaviour of its own, with its own store
reads, sitting inside a component that otherwise renders. It moves to
components/marketplaces/follow-unsubscribed.ts as useFollowUnsubscribed
and the page drops five imports with it. No baseline row: the growth
was this branch's, so the split is the answer, not a declaration.

Two other leavings of the same rebase. The table's local `repo` — the
bare repository it may be browsing — became `browsedRepo`, because
this branch gives `repo` to the subscription the page is for, and the
rename needed a line saying why. The `places` prop beside it came out
one column over biome's width.

Both were found by composing the same endpoint a second way: the
squashed branch rebased onto main in one resolution, diffed against
the eighteen picks. Nothing else differed.
