+++ ui/src/pages/marketplace-detail.tsx
@@ -1,15 +1,12 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 import type { AvailablePackage, Catalog } from "@/bindings";
 import { AboutSection } from "@/components/marketplaces/about-section";
 import { BundleCards } from "@/components/marketplaces/bundle-cards";
 import { DetailHeader } from "@/components/marketplaces/detail-header";
+import { useFollowUnsubscribed } from "@/components/marketplaces/follow-unsubscribed";
 import { MarketplacePlaces } from "@/components/marketplaces/marketplace-places";
 import { PackagesTable } from "@/components/marketplaces/packages-table";
-import {
-  marketplaceIdentity,
-  openPlace,
-  personalFirst,
-} from "@/components/marketplaces/subscribed-grouping";
+import { marketplaceIdentity } from "@/components/marketplaces/subscribed-grouping";
 import {
   useCachedRead,
   useCatalog,
@@ -24,7 +21,6 @@ import {
   catalogKey,
   marketKey,
   readErrorKey,
-  subscription,
   useMarketplacesStore,
 } from "@/stores/marketplaces";
 import { useNavStore } from "@/stores/nav";
@@ -47,9 +43,6 @@ function MarketplaceDetail({ requested }: { requested: Catalog }) {
   const rows = useMarketplacesStore((s) => s.rows);
   const packages = useMarketplacesStore((s) => s.packages);
   const load = useMarketplacesStore((s) => s.load);
-  const read = useMarketplacesStore((s) => s.read);
-  const goToMarketplace = useNavStore((s) => s.goToMarketplace);
-  const leaveMarketplace = useNavStore((s) => s.leaveMarketplace);
   const loadPackages = useMarketplacesStore((s) => s.loadPackages);
   const loadCatalogBundles = useMarketplacesStore((s) => s.loadCatalogBundles);
 
@@ -91,36 +84,7 @@ function MarketplaceDetail({ requested }: { requested: Catalog }) {
   );
   useCachedRead(bundles !== undefined, !!bundlesError, ready, readBundles);
 
-  // Unsubscribing from the Projects section can remove the very place this
-  // page opened as. The row goes, and with it the tab the person was
-  // reading — so the page follows the marketplace to another place still
-  // holding it, or leaves for the list when none does. The identity has to
-  // outlive the row to do that, and only a read that landed may be taken
-  // as proof the place is gone: rows are empty before the first read and
-  // after a failed one, and neither is an unsubscribe.
-  const lastIdentity = useRef<string | null>(null);
-  useEffect(() => {
-    if (identity) lastIdentity.current = identity;
-  }, [identity]);
-  useEffect(() => {
-    if (catalog.by !== "subscription" || row || read.status !== "landed")
-      return;
-    const held = lastIdentity.current;
-    if (!held) return;
-    // The same pick the card makes, from the same helper: a place actually
-    // offering packages, personal before a project. First-in-overview-order
-    // would land the reader on a switched-off place.
-    const elsewhere = openPlace(
-      rows.filter((r) => marketplaceIdentity(r) === held).sort(personalFirst),
-    );
-    if (elsewhere) {
-      goToMarketplace(subscription(elsewhere.scope, elsewhere.name));
-    } else {
-      // Left, not navigated away from: this page is a subscription that no
-      // longer exists, so it must not be somewhere Back can return to.
-      leaveMarketplace("subscribed");
-    }
-  }, [catalog, row, read.status, rows, goToMarketplace, leaveMarketplace]);
+  useFollowUnsubscribed(catalog, row, identity);
 
   // The tab is controlled so a section that stops existing cannot leave the
   // page with nothing selected: Projects is gone the moment the opened
